api/requirements.txt
flask
flask-restful
psycopg2-binary
Flask-SQLAlchemy
Flask-JWT-Extended
pandas
pydantic
